requirements.txt
# Voice AI Assistant - Vocode Edition
# Python 3.10+

vocode[io]
python-dotenv
pydantic-settings